Written DFA.fromLTLAutomaton(), DFA.asStringDFA()
+ fromLTLAutomaton() creates a new DFA from the automaton class of the
    ltl2automaton library. See asLTLAutomaton().
+ asStringDFA() returns a new DFA that only uses strings for
    transitions.
+ Generating Use pages with Javadoc
+ Minor changes in automata package

+++ src/automata/LatexSaver.java
@@ -30,7 +30,7 @@ public final class LatexSaver {
 	 * @return True if no errors occurred
 	 */
 	public static boolean saveLatexFile(LatexPrintableGraph graph,
-			File texFile, float spaceCm) {
+			File texFile, double spaceCm) {
 
 		// Package options
 		String classOptions = graph.standaloneClassLatexOptions();
